Clear pre-existing architecture FPs in sample-js-project-positive
The harness extension above turns on architecture-checker rules over the
positive fixture, which surfaced 26 violations already in the fixture:

  15x architecture/deterministic/dead-method
   5x architecture/deterministic/unused-export
   3x architecture/deterministic/god-module
   3x architecture/deterministic/dead-module

Cleared by fixture changes only — no visitor changes:

- shared/utils/src/logger.ts: convert object-literal arrows to top-level
  `export function` declarations and re-bind them on a `logger` const so
  every `logger.info(...)` call site keeps working. With the named
  exports the dead-method skip recognizes cross-file `logger.info`/
  `.warn`/`.error` calls.
- shared/utils/src/profile.ts (new): consumes formatters/validators/
  logger directly, providing a true (non-barrel) module-level dependency
  to formatters.ts and validators.ts so they're no longer dead modules.
- shared/utils/src/index.ts: re-exports `describeUser`/`ProfileInput`,
  with the type re-exported via `export type { ... }` to satisfy the
  mixed-type-exports rule.
- shared/utils/src/code-quality-{advanced,misc}.ts and legacy.js: split
  each into two files (`*-extras.{ts,js}`) so neither side exceeds the
  god-module threshold of 15 methods. Function semantics unchanged.
- services/api-gateway/src/services/wiring.ts (new): composes the
  gateway's services. Instantiates UserService and HealthService and
  calls each method; calls `checkLimit`, `getConstants`, the registry's
  job runner via the new exported `notifyUserJobRun` shim, and
  `describeUser`. The gateway's `startApp` invokes `describeGateway`
  so the composition is reachable.
- services/api-gateway/src/index.ts: imports + calls `describeGateway`.
- services/api-gateway/src/jobs/notification-job-from-documenso-documenso.ts:
  extract the inline `run` arrow to an exported `notifyUserJobRun`
  function, aliased back into the object as `run`. Lets the dead-method
  skip see the direct cross-file call.
- services/user-service/src/index.ts: imports + calls `getStatusCodes`
  and `validateEmail` from `./handlers/user.handler`.
- tests/fixtures/sample-js-project-positive/externals.d.ts: declare the
  newly-exposed `@sample/shared-utils` exports (`validateName`,
  `formatDate`, `ProfileInput`, `describeUser`) so the fixture
  type-checks.

Full suite green: 58 files, 3454 tests.
