+++ tests/fixtures/sample-js-project-positive/services/api-gateway/src/index.ts
@@ -1,7 +1,10 @@
 import { authMiddleware } from './middleware/auth';
 import { logger } from '@sample/shared-utils';
+import { describeGateway } from './services/wiring';
 export function startApp(): void {
   logger.info(`Auth: ${typeof authMiddleware}`);
+  const summary = describeGateway('127.0.0.1');
+  logger.info(`Gateway ready: ${summary.userIndex}`);
 }
 process.on('uncaughtException', (err: Error) => {
   console.error(err.message);
