+++ tests/fixtures/sample-js-project-positive/shared/utils/src/code-quality-advanced.ts
@@ -30,97 +30,3 @@ export function mapItems(items: readonly string[]): string[] {
 // positive: ts-declaration-style — interface with extends clause should NOT trigger
 interface BaseConfig { timeout: number; }
 export interface ExtendedConfig extends BaseConfig { retries: number; }
-
-// positive: insecure-random — Math.random() for non-security array index should NOT trigger
-export function getRandomElement<T>(items: readonly T[]): T {
-  return items[Math.floor(Math.random() * items.length)];
-}
-
-// Positive: unused-constructor-result — new URL for validation
-export function validateUrl(input: string): boolean { try { new URL(input); return true; } catch { return false; } }
-
-// Positive: complex-type-alias — simple string literal union
-export type Status = 'active' | 'inactive' | 'pending' | 'archived' | 'deleted';
-
-// Positive: indexed-loop-over-for-of — partial range loop (needs index arithmetic)
-export function pairwise(items: readonly number[]): number { let sum = 0; for (let i = 0; i < items.length - 1; i++) { sum += items[i] + items[i + 1]; } return sum; }
-
-// Positive: unused-collection — collection reassigned and returned
-export function buildList(): string[] { let items: string[] = ['a']; items = [...items, 'b']; return items; }
-
-// Positive: json-parse-in-loop — parsing different strings each iteration (not same string)
-export function parseAll(items: readonly string[]): unknown[] {
-  const results: unknown[] = [];
-  for (const item of items) {
-    try {
-      results.push(JSON.parse(item));
-    } catch {
-      // skip invalid JSON
-    }
-  }
-  return results;
-}
-
-// Positive: prototype-pollution — Object.entries iteration (safe, not bracket assignment from user input)
-export function applyMapping(target: Record<string, string>, source: Record<string, string>): void {
-  for (const [key, val] of Object.entries(source)) {
-    target[key] = val;
-  }
-}
-
-// Positive: missing-react-memo — TS generics with comparison operators must NOT be flagged as JSX.
-// Pre-fix the rule used `text.includes('<') && text.includes('>')` and matched generics.
-export function compareGeneric<T extends number>(a: T, b: T): number {
-  if (a > b) return 1;
-  if (a < b) return -1;
-  return 0;
-}
-
-// Positive: missing-react-memo — nested generics in return type, no JSX
-export function pairBuilder<K, V>(key: K, value: V): Map<K, V> {
-  const map = new Map<K, V>();
-  map.set(key, value);
-  return map;
-}
-
-// Positive: floating-promise — sync functions whose names match old ASYNC_PREFIXES.
-// Pre-fix these were flagged because the rule used a name heuristic.
-// Phase 4 (TypeQueryService) verifies via real type info that these return non-Promise.
-const DEFAULT_SIZE = 1024;
-export function createBufferSync(size: number): ArrayBuffer {
-  return new ArrayBuffer(size);
-}
-export function loadConfigSync(): { ready: boolean } {
-  return { ready: true };
-}
-export function callSyncCreate(): ArrayBuffer {
-  // Bare call of a sync function whose name starts with "create" — pre-fix the
-  // ASYNC_PREFIXES heuristic flagged this. Type info now correctly skips it.
-  return createBufferSync(DEFAULT_SIZE);
-}
-
-// ---------------------------------------------------------------------------
-// Phase 3: framework overfit fixes
-// ---------------------------------------------------------------------------
-
-// Positive: state-update-in-loop — bare setX() identifier call in a loop in a
-// non-React file. Pre-fix the rule fired on any /^set[A-Z]/ call in a loop;
-// now gated by React import (this file does not import React).
-declare const setStringValue: (s: string) => void;
-export function applyTitles(items: readonly string[]): void {
-  for (const item of items) {
-    setStringValue(item);
-  }
-}
-
-// Positive: static-method-candidate — Vue lifecycle method on a class extending
-// a Vue base. Pre-fix the hardcoded React-only contractMethods list missed
-// Vue/Angular/Svelte lifecycle methods. Now: classes that extend ANY base are
-// skipped via the existing heritage check.
-declare const VueBase: { new(): { mounted(): void } };
-class MyVueComponent extends VueBase {
-  mounted(): void {
-    console.warn('mounted');
-  }
-}
-export const vueComponent = new MyVueComponent();
