+++ tests/fixtures/sample-js-project-positive/shared/utils/src/logger.ts
@@ -1,11 +1,16 @@
-export const logger = {
-  info: (message: string): void => {
-    process.stdout.write(`[INFO] ${message}\n`);
-  },
-  error: (message: string): void => {
-    process.stderr.write(`[ERROR] ${message}\n`);
-  },
-  warn: (message: string): void => {
-    process.stderr.write(`[WARN] ${message}\n`);
-  },
-};
+export function info(message: string): void {
+  process.stdout.write(`[INFO] ${message}\n`);
+}
+
+export function error(message: string): void {
+  process.stderr.write(`[ERROR] ${message}\n`);
+}
+
+export function warn(message: string): void {
+  process.stderr.write(`[WARN] ${message}\n`);
+}
+
+// Named exports are the canonical entrypoint; the `logger` object is a
+// convenience alias kept so existing `logger.info(...)` call sites keep
+// working without churn.
+export const logger = { info, error, warn };
